refactor: converted all names ot follow snake_case
