+++ app/dashboard/events/page.tsx
@@ -1,5 +1,5 @@
 
-import EventsTable from "@/components/app/EventsTable";
+import EventsTable from "@/components/app/events_table";
 import { fetchAllOrganizations, fetchOrganizationsForUser } from "@/lib/organization";
 import { createClient, getUser } from "@/lib/supabase/server";
 import { Event, Organization } from "@/lib/types";
