+++ app/e/[slug]/page.tsx
@@ -1,6 +1,6 @@
 "use client";
-import Footer from "@/components/Footer";
-import Header from "@/components/Header";
+import Footer from "@/components/footer";
+import Header from "@/components/header";
 import Preloader from "@/components/preloader";
 import {
   checkEventPrivacyAndMembership,
