+++ components/app/organization_card.tsx
