+++ components/combined_user_data_table.tsx
