+++ components/dashboard/statistics_section.tsx
