+++ components/events_table.tsx
