+++ components/pricing_section.tsx
